requirements.txt
torch>=0.4.1
pillow
numpy
torchvision>=0.2.5
scikit-learn
